Create a performance testing script.
This script will let us measure the performance (build time, execution
time, and binary size) of generated protos and the runtime library using
a message with various numbers and types of fields (provided via the
command line).

It builds and runs against the runtime and plugin in the parent
directory, so local changes can be easily tested and compared.