Restrict visibility of some types to internal.
Namely, the encoding visitors can be internal since they are only used
by extensions within the runtime.

There's more work to be done here; hopefully more parts of the decoders
can be internalized as well, but there are protocol/generic constraint
relationships that need to be pulled apart first. That will have to
happen as part of the larger segregation of JSON/text from binary
effort.