Remove [UInt8]-based serialization methods.
The performance difference between these and `Data` appears to be noise
(which makes sense; they both just grab the underlying pointer and work
on it). Since `Data` is the standard Foundation data type for this kind
of raw data and it's what most users will use when working with files
and streams, we'll keep these and remove the array-based ones.

This also lets us keep a consistent interface when we improve
performance for discontiguous `Data` later on.

Fixes #44.