Fix linker errors in release builds. (#71)
When using the `ProtobufWrapper.init(_ value: WrappedType.BaseType`)
initializer from externally linked code, the linker dies in release
builds with errors that it cannot find the protocol witness table for
the initializers on the concrete types. This *may* be a Swift bug;
either the compiler should not allow the mismatched visibility on the
conformances, or the linker should allow the call.

For now we duplicate the implementation throughout the concrete types
instead, despite the small amount of bloat this adds.

Fixes #70.