Issues #127, #128, #129:  Rework unknown field handling (part 1)
Add some tests that clearly document the behavior I'm trying to
satisfy.  Specifically, proto2 guarantees that packed fields behave
the same as repeated fields across decode/recode/decode cycles.  If
you have regular repeated enum fields, the specific fields that
contain unrecognized values get pushed back as "unknown" and
reserialized at the end.  This changes the order on a subsequent
decode, but preserves all of the original values.

These tests just verify the above behavior by deserializing
repeated and packed enum fields with a mix of known and
unknown values and checking that a subsequent recode preserves
all the values (but with the unrecognized values pushed back
to the end).