Extension naming/scoping changes.
The naming/scoping of an extension depends on where the extend {} is declared
(top level or within a message). The message being extended shouldn't factor
in to the naming of the extension. This is the model used by other languages
and allows two files to extend a common message with the same name as long as
the field numbers are different. What is not allowed in two extensions defined
at the same scope with the same name but on different messages (again, the
extendee isn't part of the naming).

- Top level extend {} blocks should get the fields and extension they add
  prefixed with the defining proto package (or Swift prefix).
- For top level extend {} usage; add "Extensions_" between the proto package
  and the extension name. This is like the struct Extensions used in message
  scoped extend {} blocks. This gives the extension definition (the `let`) a
  different name from the Swift properties added via the Swift extension so
  the generated code for them can reference the defined extension.
- An extend {} scoped within a message, gets named with the parent message
  scoping it.
- Update the unittesting code for the naming changes.

Fixes https://github.com/apple/swift-protobuf/issues/96