Stop special casing defaults for equality
Hash and Equality need to be in sync; so for proto2 syntax, equality is about
what fields are set and set to the same thing, so defaults shouldn't be
special cased.

While the code gets a little larger, it actually should be a little faster also
as it means the generated code is calling the private values and isn't calling
the helper where returns the default to then compare that.

Disable some required fields tests; the current behavior of required fields
is wrong in they it doesn't check require them to be set and instead just
always serializes them, with the change to equality, they test no longer
pass but the fix is fixing up required handing in general.