Don't let Instruments interfere with timing results
Instruments has a lot of overhead and introduces additional
variability into the timing.  Instead, run the harness
once to collect timing data (without using Instruments and
waiting a couple of seconds before and after for the system
to settle) and then a second time under Instruments (to collect
memory use metrics).