Streamline the binary decoding path for fixed-size values
This reworks the fixed4 and fixed8 decoding to pass down
a scanner reference and have the type use that directly, rather
than pre-parsing (and copying) the field contents.

This also eliminates a lot of `UnsafeBufferPointer` structs
in favor of just passing the pointer/count as separate arguments.

This seems to provide about a 4x speedup for the affected data
types.