Speed up the JSON decoder
This is a pretty thorough reimplementation of the JSON decoder that
adopts some of the recent changes made to the Text and Protobuf Binary
decoders.  In particular, it entirely abandons the old token-based
model -- parsing is now lazy and schema-driven, and skipping (for
unknown fields and Any) involves text-level parsing rather than
token-level.

Decoding now involves a single JSONDecoder object that holds a single
scanner object -- these are both created at the top level and passed
down as an `inout` parameter through nested calls.  Among other
things, this pattern minimizes refcounting overhead.

Numeric parsing now uses the same logic whether the number is quoted
or not.  The top-level code looks at the first character; if it's a
quote, it skips the quote, parses the number, and then verifies a
trailing quote after the number.  In particular, we usually entirely
avoid creating an intermediate string.  One odd corner case: if the
number is in quotes and uses backslash escapes, then we fall back to a
slow path which parses and decodes the string then re-parses the
contents as a number.  Of course, this path will probably never be
exercised outside of tests.

Any now stores a `[String:String]` with the JSON text of the
deferred fields instead of a `[String:[JSONToken]]` holding
lists of tokens for each deferred field.  I haven't measured
performance for this, but I expect it will be noticably faster.

The total speedup is around 2x for the field types I've tested.  This
further improves our performance advantage over Google's C++ protobuf
but we're still noticably slower than dedicated JSON implementations
such as Foundation's JSONSerialization API.

There are still areas where performance could be improved:

 * Getting the next field number still involves creating a temporary
   string that we can look up in a hash table.  Creating this temp
   string can account for 50% of the total run time, so alternate
   lookup strategies could be a big win.

 * As much as 1/4 of the total time is still spent in refcounting
   overhead.  I'm still trying to understand where this is coming
   from.

 * The three decoders are much more similar in structure now; it may
   be possible to combine some of the code.