Print times for 1 operation, use tabular output, skip C++ Instruments
This changes the timing printout to divide by the run_count, effectively
printing the time for a single operation.  This also required changing the
units to microseconds instead of milliseconds.

This makes it feasible to change the run_count and get comparable results,
which is useful as things get faster and you want to bump up the run_count
to get better precision.  I've gone ahead and bumped the run_count to 1000
while I'm here.

I've also reworked the display during the run to use a tabular format
and reordered the overall flow to first run C++ alone, then Swift alone,
and finally run Swift under Instruments.  This allows you to look at and
think about the values while waiting on Instruments.

I've also shuffled the shell script a bit to omit C++ under Instruments;
it's easy to put that back if you really want to.