Clean up error handling per Issue #205
This splits the decoding errors into separate enums for:
 * JSONDecodingError
 * TextDecodingError
 * ProtobufDecodingError
 * AnyUnpackError

Since we no longer have a generic "DecodingError", I've
changed how OneOf conflicts are handled so that the decoder
is the one that actually throws the error.  This also divorces
the generated code from the error type details, which is a nice
code cleanup.