Issue #222: Flag Map keys during JSON parsing so we can distinguish "true" from true.
Google's conformance tester flagged us for accepting quoted "true"
as a synonym for unquoted true everywhere.  We now set a flag on
the JSONDecoder struct when decoding a map key so the decoders
can distinguish these cases (it happens that bool is the only
type for which the quoted form is not accepted everywhere).