Make extension field name handling private.
Specifically:
 * Drop _protobuf_fieldName entirely
 * Make the extension field name a simple
   StaticString in the generated code
 * Drop the rest of ProtoNameResolver
   (we've essentially inlined the last of that
   directly into TextFormatEncodingVisitor)

This hides the details of extension field name handling inside the
library, and makes it an implementation detail shared between the
extension types and the TextFormat encoder.

This required making the _extensionFieldValues set public
so that the TextFormat can get the object and use it directly
for name lookups.

As a side effect, this also allows us to move the duplicated
{get,set,has,clear}ExtensionValue boilerplate into the
ExtensibleMessage protocol, further simplifying the generated code.