Compute JSON field names at run time
The generator now test-converts each proto name to JSON.  If the JSON
can be derived from the proto name, the generator stores only the
proto name and the library will convert at run-time.  In the library,
this uses an interned string pool to cache the UTF8 so there should be
no run-time performance penalty compared to previously (just some
additional memory to track whether the UTF8 comes from a StaticString
or from the cache).