Use numeric field names if the name isn't available
In particular, we should be able to build text format output
regardless of whether names are available.

To implement this, I've refactored a bit, removing the optimization
that old-style repeated fields only look up the name once.  Name
lookup in general is a lot faster than when I put that in, so it may
not be needed.  (If it is, we can revisit.)