Derive the reserved word lists from the Swift language spec
This adds collections of Swift reserved words to
PluginLibrary/SwiftLanguage that are taken directly from
the language spec.  It then redefines the reserved word
lists used to filter struct, enum, and field names in terms
of these.

This should clarify why various words are reserved by the
code generator and hopefully make it easier to maintain
these lists over time.  In particular, the lists of Swift
reserved words can just be copied/updated from the langauge
spec as it evolves.