Don't add spaces before the comments pulled from proto file.
protoc captures the leading spaces on the proto file comments; so
never add spaces between the "//" or "///" the generator puts
on the front and instead rely on the spacing from the proto file.

When looking at the generated source, it seemed overly the content
after the comment marker was overly indented; this seems (to me)
to be a little more natural.