Issue #40: Expanded testing of magic field/enum/message names
This mines the SwiftProtobuf library source code to generate
a mined_words.txt list of ~600 Swift names that might potentially
cause problems.  This list includes Swift type names and reserved
words as well as identifiers used by the library that might
conflict.

This list is used to generate several proto files that use these words
as the names of messages, fields, enums, and enum cases.  These proto
files are rebuilt as part of `make regenerate` so they should remain
up-to-date over time.

Based on this test, add several names to our ReservedWords list
and update the sanitizing logic to better handle certain cases.