Fixed oneof isInitialized generation when all fields are messages with required fields (#461)
Previously oneOf was generated with case "None" and oneOf var was non-optional.
Currently, it's generated without "None" case, but optional var, so when isInitialized is generated and all cases in oneOf are messages and have required fields, "default" is not generated which leads to compiler error "Switch must be exhaustive" (nil case is not covered). Looks like having oneOf as optional var requires "default" to be generated always.