Add descriptor objects to the PluginLibrary.
These provides the same sorta info a google/protobuf does with
their src/google/protobuf/descriptor.{h,cc} -- a graph of the
data in the protos so a generator can walk up/down, and directly
to resolved info.

Add unittests to ensure the graph is built and resolved as expected.