Add a SwiftProtobufNamer to encapsulate the generation of the library names.
Keep the Descriptor object to be just models of what comes in from protoc
and instead use the Namer as something that can generate the library specific
names. This should help avoid the Descriptors becoming to close to what the
generators in a generator would be, but still allow providing the Namer as a
building block for anything that builds on top of SwiftProtobuf.

Also move unknown enum preserving semantics helpers to the generator level
(for the same reasoning).