Generate a traverse for each subgroup.
When the oneof isn't continuous in parent message, generate a different
traverse method for each "group", this avoid the need for conditionals
within the cases.

Name the traverses based off the field number(s) in them so reading of
the message traversal is a little more clear also.

If a traversal switch would only have one case, use the if case syntax
instead.