Regenerate the files.
The generator had been coded to properly always generate "SwiftProtobuf." on
all the types the library uses to function to help sure there aren't conflicts
with anything a developer defines; but the missing part was any place a proto
referenced one of the generated types from the bundled protos (the Well Known
Types). With the plumbing in place for the modules, this now auto inserts the
"SwiftProtobuf." on all those types since they are auto added to the mapper.