Performance: Speed up decoding of TextFormat `bytes` fields by ~4x
This eliminates constructing an intermediate string
and instead parses the UTF8 bytes directly into
a Data.  To avoid expensive resize operations, we
parse the field twice:  The first time we validate
the text format (including backslash escapes) and
count the number of bytes that will be generated;
then we create a Data() object of the right size
and decode directly into it.