Sync up the merging decode and decoding initializer.
When decoding repeated fields, there isn't really a good way to tell
if something decoded.  The types that support packing also add to
the difficulties in that a zero length data blob is "valid" in that
it can cleanly decode, but it really shouldn't happen. So...

Make merging decode and initialization decode act the same in that
neither ensure something was added to the internal array.