Tighten up the bool TextFormat parsing.
During TextFormat parsing, if the field was boolean, only the first
character was used if it was "0" or "1", not the next whole token.
This let bad input either choke some what by accident later or
could allow things to slip through as valid when no other text
parser could likely parse it.