Add a BinaryDecodingOption to discard unknown fields.
As part of the change to proto3 syntax, some other languages are
providing means to drop unknowns while parsing and/or an api to
strip unknowns from the object graph. Stripping from the graph
would require generating a new method (since we don't have
reflection), but it is pretty easy to add an option now to drop
the unknowns during parsing.

- While at it, tweak the repeated enum parsing to not create the array
  until it is needed (since it likely isn't needed the majority of the
  time).
- Add tests to cover the option.