Group the Swift Extensions for Proto Extensions.
Instead of making a Swift Extension for every Proto Extension, pre order
the list of Proto Extensions so we can generate one Swift Extension to
the subset of Proto Extensions.

This helps make the generated code a little easier to navigate as well as
a little more like raw proto files where an "extend MyMessage" can declare
multiple fields; and the generated code is now in that pattern.