Add Swift 4 string decoder
This takes advantage of the (much faster) `String(decoding:as:)`
initializer in the current Swift 4 development snapshots to improve
String decoding speed.  This has big performance implications for
binary, JSON, and text decoders as well as JSON and text encoding.

I've also reorganized this file to try to make it easier to
follow.