Avoid extra buffering when writing out sub messages.
Instead of creating a complete Data for the submessage, compute the
length (which would have been done anyway inside of serializedData()),
write out that varint, and then traverse the message to write its
content.  This avoids the allocation of the Data for the sub message
and then having to copy it into the real Data being created.