Add a helper for "delimited message" reading/writing.
The Java implementation long ago added the concept of a "delimited message" when
reading/writing to a stream. Normally, you can concatenate the serialized form
together and it is still one message; but this form used a varint to write
the length and then write the message, allowing multiple messages to be
written distinctly.

In github.com/google/protobuf/pull/710 Google's protobuf team has
suggested generally supporting this, but the think it should be as
utilities rather than directly on the Message classes. The logic
being there could be multiple ways to "frame" messages for different
transports, and adding them all to Message can lead to bloat/confusion
on the over all api surface.

For these reasons, the api here is static functions scoped to an enum to
provide logical grouping, but not pollute the Message's namespace. Swift
extensions on InputStream/OutputStream could also be done, but since lots
of uses of protocol buffers might never need this, it seemed better not to
pollute those namespaces either.