Use DoubleFormatter.utf8ToDouble() for parsing doubles.
The crux of this change is just to add a DoubleFormatter instance
to JSONScanner and move `parseBareFloatString()` into JSONScanner
so it can use that object for the parsing.  (TextFormat already
worked this way.)

Along the way, I changed the name to `parseBareDouble()` to reflect
that it now returns a `Double?` instead of a `String?`, and simplified
a bunch of places that can now avoid juggling the no-longer-needed
String intermediate.

This is a significant performance boost when parsing JSON objects that
contain float or double fields: `perf_runner 100 double`
is almost 3x faster with this change.