Swift 3 and Swift 4 have different names for narrowing integer conversions
Since we use those so widely, put a layer of indirection into
MathUtils.swift (lowercase 'int(truncating:)', etc) so we can adapt
between Swift 3 and 4 in just one place.

I propose that when Swift 5 comes out, we drop Swift 3 support.
Hopefully, at that time, we can just inline the Swift 4+ version
of this code and lose the indirection.