Fix ambiguous reference to encoder struct
The new Swift compiler complains (correctly) that the
reference to self.encoder here was ambiguous; technically,
self and self.encoder are both value types that are captured
(copied) by the closure before `_visitPacked` is called.
That would strictly require the encoding operations here to
be discarded (since they are getting invoked on a copy of self
that is then discarded).

That was clearly not the intention of this code, which is
why the new compiler is complaining.

The fix here is to change the closure to accept a mutable reference to
the encoder, making the intention explicit.