Regenerate proto source and update code.
google/protobuf/test_messages_proto3.proto had the message
name tweaked when adding proto2 conformance tests. For other
languages, the file is specific to the conformance tests, but
SwiftProtobuf is using it for a bunch of unittests also, so those
tests also had to be updated for the new naming.