Perf: Rework Text and JSON enum handling to take full advantage of NameMap.
On decode, we were converting the UTF-8 enum name into a String and
then NameMap was converting it back to UTF-8 for lookup.  The code now
short-circuits that, looking up the UTF-8 key directly on the NameMap.

For JSON, a quoted enum key is allowed to have backslash escapes.
This should be very rare since proto enum names are always simple
ASCII, so we can optimize by first trying a fast path that returns
UTF-8 for a key with no such escapes, then falling back to a slower
String decode if it must.

On encode, JSON repeated fields (including enums) were spending a lot
of unnecessary time converting constant strings such as "[" and ","
to/from UTF8.  This commit fixes that.

I've also edited the Performance harness to support `enum` as a field
type.