Perf harness: Populate repeated bool fields consistently
In particular, this fixes the discrepancy between the number
of items in a `repeated bool` array between C++ and Swift
test harnesses.

Note:  "Populate Fields" for `100 'repeated bool'` on my
laptop went from 13us (old version that just assigned a full array)
to 233us (new version that called `.append` in a loop).
Compare this with ~40us for the C++ version.