Tweak to oneofs on Descriptors.
The oneofIndex is used as the more common test and happens as part of
bind anyways, so do it during init and cache the result. Instead do
the descriptor lookup for the oneof on demand since it isn't used
nearly as often.

This means the lookup only has to happen when generating a file and
not when building the the descriptor graph for imported files which
will never care about the oneofs in other files.