Move the SwiftProtobuf.Message conformance into the extension.
Swift 4.1 is adding "near miss" warning to catch typos around overrides for
protocols items that have default implementations. This can fire on the
generated code for things like "hasValue" vs "hashValue" or "isInitialized_p"
vs "isInitialized".

There is no way to suppress these checks, and even if they were more strict
and took the return types into account, they could still happen. The library
completely avoids the problem space by moving the declaration of
SwiftProtobuf.Message conformance to an extension so the body of the extension
doesn't appear to have any of these near misses via the message's fields.

While at it, the methods needed for SwiftProtobuf.Message conformance are also
moved into the extensions, so a developer jumping to the type in the generated
source won't also be presented with as many internal details of the library.

Fixes #719