Issue #743: Correctly handle escaped slashes in JSON byte data
Protobuf JSON is designed to be easily implemented on top of standard
JSON encoders.  For historical reasons, such encoders often encode "/"
(which commonly occurs in base-64 data) as "\/".

I overlooked support for backslash escapes when parsing byte data.

This adds tests and support for all the simple backslash escapes.
It turns out that "\/" is the only one that is legal in base-64 data,
so we simply fail if we encounter \b \f \n \r \t \" \\ or any invalid
sequence.

Note: This does not support \u1234 hexadecimal escapes.  (I don't
expect to ever see those in practice, so I doubt this is a problem.)