Extend the Vistor protocol to express Hashable on Messages.
In prep for better support for SE0206, extend Vistor to require
the Messages also be Hashable. This will let us remove custom code
around the collections and let the newer collection code directly
do the hashing.

This ripples to the Extension support where some of those types
also need to move of `Message & Hashable` so they can the use
the Vistor apis.