Remove reference to api that are "hookable".
These fall down in the face of modules. What's worse, is they
don't fail to build, the fail to apply in some cases, and to
understand it, a developer has to really understand how protocols
work/are resolved/etc.

We also were documenting the wrong method as a hook point.

So, rather than try to document all the things that do/don't work,
just stop documenting it, and allow future changes to not be
required to support it.

Issue #773 is open to look at provide better apis in the future for
customizing comparisons.