Use access modifiers directly on decls instead of on extensions.
Code was a mix of modifiers on decls or extensions, which meant figuring
out the access when looking at the code was a mix of seeing it and having
to scroll back to the access on the extension to check.

Swift 5 dev snapshots also kick out a warning for when the modifiers are
redundant between the extension and decls, so by removing them from the
extension it avoids those warnings.