Performance: Simplify formatting double/float in JSON and TextFormat
Previously, this code directly invoked C standard
library `sprintf` to convert floating-point numbers to text
formats.  The new code uses Swift standard library routines
that are both faster and more accurate (in Swift 4.2 and later).
In particular, most of `DoubleFormatter` has been removed
and the remainder renamed to to `DoubleParser`.

Swift 4.2 and later:

New versions of Swift have floating-point formatting
logic that is both fast and always generates "optimal"
output that is guaranteed to parse back to exactly the
same value.  Note that `description` and `debugDescription`
generate the exact same output for all finite values.

In particular, the following is now a fast way to
append the UTF-8 form of a floating point number to
a Data object:
```
  data.append(contentsOf: value.debugDescription.utf8)
```

Swift 4.1 and earlier:

Correctness:  The snippet above uses `debugDescription`
since Swift 4.1 and earlier did not always produce enough
digits for `description` to ensure round-trip accuracy.

Performance:  The new code is a modest pessimization in Swift 4.1
and earlier; the C library was used previously because Swift's
standard library routines for floating-point formatting were not
very fast.