Revert "Drop Swift 4.0 support (#905)"
This reverts commit 687cff576a4a9b52cb34cf6b70766e996525b522.

SwiftPM only has constants for 4.0 and 4.2 (as well as the compiler only taking
those values (and not 4.1)). So to drop the 4.0 from there, it would actually
mean moving up to 4.2 which is more agressive than our policy, so it seem like
we have to keep 4.0 support because you can't tell the SwiftPM or the compiler
to be atleast 4.1. So we can't drop support until we can move all the way up
to 4.2.