Use raw pointers instead of typed pointers.
In general, it is better to use raw pointers instead of typed pointers
whenever those pointers are provided to you by external users. The
reason for this is that Swift requires that each pointer be bound to one
and only one data type at any point in the program execution. If you are
vended a raw pointer (e.g. from a Data), your program does not
unequivocally know what the type of that pointer may be, and so to
assume that it is bound to UInt8 or Int8 is to risk violating Swift's
pointer aliasing rules.

We've learned lessons in the past that violating the pointer aliasing
rules in programming languages can lead to disaster. In our case, we
can avoid the majority of this trouble by electing to use raw pointers
instead of typed pointers. This also helps us by removing tricky pointer
management code.

This patch contains essentially no function change: wherever possible I
made straightforward, targeted refactors. I think once or twice I turned
a pointer into a buffer pointer, but I generally resisted getting too
frisky with changes.