Guard against stack overflow when generating TextFormat for unknowns.
The TextFormat encoding attempts to detect when a length delimited field appears
to be a message and then decode it as such. This means a well crafted message
could be fed to a process in binary format to attack it if one knows in some
condition it will generate the TextFormat (say for logging). This avoids the
problem by limiting the recursion when doing this additional decoding.