Use dictionary subscript with default.
https://developer.apple.com/documentation/swift/dictionary/2894528-subscript
says it goes back to Xcode 9, so at this point, our min Swift version should be
ensuring things always support it.

The only other place that it might make sense to use this appears to be within
the SimpleExtensionMap, but the values looked up get filtered before adding to
them, so it seems like the fast path might for not being in the map might still
be worth while.