Expand JSON test helpers to also verify the UTF-8 data methods
This will help ensure that both work exactly the same.  There
have been a couple of minor bugs recently where preconditions
were being checked differently in the two paths.

While here, I noticed that we were not correctly verifying that
decoded JSON was correctly re-encoded.

Fixing that in turn revealed a test bug:  I was asserting that
signed zeros were always preserved, but they are currently not.
By my reading of the spec, because -0.0 == +0.0, signed zeros
should be treated as default zero values and be omitted.