Reflow the camelcase apis so they can support non identifier input.
The one place Swift identifiers are made from non protobuf identifiers is for
the extension registry created for files with extensions. It uses the filename
in the creation of the identifier, and based on the file name it could result in
and invalid Swift symbol. This changes the camel casing code to handle things
outside the protobuf identifier range to ensure the end up as valid Swift
symbols also.

While the support for non protobuf identifier characters is new in this code,
the reality is it would have resulted in invalid Swift, so it wasn't possible
before. i.e. - what didn't work, now works. And the mapping for filenames that
were using valid protobuf identifier characters remains unchanged, so no
breaking behavior.

- Rather than splitting and reassembling, build up the transformed result as
  things go.
- Use an enum and switch to make the camelcasing flow clearer.
- Support non proto identifier characters:
  - if valid Swift identifier characters, take them as is; otherwise bring them
    over as "_u####", this is similar to what C++ does for crafting their
    identifiers.
- Document the apis that are exposed.
- Remove the temp api used for extensions now that non proto identifiers are
  also handled.
- Add tests for the non proto identifier support.