Add test for non protobuf identifier handling.
This is just a compile test.

Tests/SwiftProtobufPluginLibraryTests/Test_NamingUtils.swift also covers this
behavior but making a full compile test to be safe.

NOTE: This is a distict commit incase git/filesystem have issues with this file
name and it ends up not working on linux. If that happens the commit can be
reverted to remove this without losing any of the supporting code.