Revert "Add test for non protobuf identifier handling."
This reverts commit 8d80630516e797f62a6fa41174e5bc6694832eed.

I've seen this trip up in one case already, so going to just remove it. As
mentioned before, there is unittest coverage, this was more like integration
testing, but isn't worth the potential issues since source and live on lots of
file systems where bad things can happen with encodings.