Add specialized defaults for repeated fields.
Setting a repeated extension field to an empty array would result it that value
being stored, so has* came back at true even though there was nothing to include
in binary serialization.

In addition to fixing that bug be behavior, this clears the way for eventually
deprecating/removing the generation of has*/clear* for repeated extension
properties.

Progress on #944