Support proto3 optional in generation.
- Filter out synthetic oneofs from oneof generation
- Map proto3 optional fields to normal fields (ignoring the oneof).
- Use the new api on descriptor to tell if a field gets presence support.
- Regenerate the files with the support.

Note: Yes, the diffs in the generated show yet another issue with our naming,
the wrapper oneof (before we filtered it out) was a naming collision with the
field. Updating our naming support to consider all other fields/oneofs for
collisions is something we've never done, but might actually have to happen
since we do transform names. Until then, issue like this will remain things that
the compiler catchs and the generator never works around.