Fix JSON coding/decoding of NullValue WKT
The google.protobuf.NullValue well-known type is used
to represent JSON `null` values.  As such, it needs custom
encoding/decoding in order to parse itself from a `null`
value or to write itself out in that form.

Prior to this, Message types could customize their JSON
encoding/decoding by conforming to _CustomJSONCodable.
This PR allows Enum types to do the same:
* They can choose to decode themselves from a `null` token
  (the null token is otherwise treated as a missing value)
* They can provide custom encoding (otherwise, the
  enum value is written out as a string or int)

It then uses this to provide the necessary customization for
the NullValue type.  Plus tests, of course.

Thanks to the Google protobuf team's ongoing conformance test
expansion efforts for catching this oversight!