Correctly handle repeated NullValue fields
Just as NullValue parses `null` and serializes as `null` when in a
singular field, it should do the same when NullValue is in a repeated
field.

This requires a new test proto (since there are no `repeated
NullValue` fields in the existing test protos) and tests to verify it,
as well as the serialization/deserialization support.