Tune the CI builds of protobuf
Build the explicit targets need instead of the default target (usually "all")

Add -j CPU count to the builds to build some things in parallel. Make doesn't
have a way to say scale to the CPU count, so look up the CPU count and
explicitly pass it.  The agents on github seem to all be dual core, so this
should help a little.